+++ Concepts/DocRetrieval/src/main.py
@@ -1,4 +1,4 @@
-from config import SITE_CONFIG
+from Concepts.DocRetrieval.src.config import SITE_CONFIG
 from scraper import bfs_scrape
 
 def main():
